chore: Add additional tests for double separators
Because we now add a trailing slash, users using the previous library version could potentially be adding an unnecessary trailing slash. But the Java File, and its Android implementation, protect against that by default, as illustrated by the the changes in the tests, which continue to pass.

References: https://outsystemsrd.atlassian.net/browse/RMET-4221